+++ tools/check-enforcer/Azure.Sdk.Tools.CheckEnforcer/IConfigurationStore.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Azure.Sdk.Tools.CheckEnforcer
+{
+    public interface IConfigurationStore
+    {
+        int ApplicationID { get; }
+        string ApplicationName { get; }
+        int ApplicationTokenLifetimeInMinutes { get; }
+
+        int MinimumCheckRuns { get; }
+    }
+}
